jwt 인증 방식 보안

+++ senials_frontend/src/pages/hobby/HobbyReviewModify.js
@@ -4,6 +4,7 @@ import { useNavigate,useLocation } from 'react-router-dom';
 import axios from 'axios';
 import { setHobbyReview } from '../../redux/hobbySlice';
 import { useSelector,useDispatch } from 'react-redux';
+import { getStoredToken } from '../../utils/authToken';
 
 function HobbyReviewGet() {
     const dispatch=useDispatch();
@@ -24,10 +25,14 @@ function HobbyReviewGet() {
 
   // 기존 데이터 조회
   useEffect(() => {
-      const token = localStorage.getItem("token");
+      const token = getStoredToken();
+      if (!token) {
+          navigate('/login');
+          return;
+      }
       axios.get(`/${hobbyNumber}/hobby-review/${reviewNumber}`,{
           headers: {
-              'Authorization': token // JWT 토큰을 Authorization 헤더에 추가
+              'Authorization': `Bearer ${token}` // JWT 토큰을 Authorization 헤더에 추가
           }
       })
         .then(response => {
@@ -70,10 +75,14 @@ function HobbyReviewGet() {
     //해당 리뷰 삭제 이벤트
     const linkDeleteReview = async () => {
         try {
-            const token = localStorage.getItem("token");
+            const token = getStoredToken();
+            if (!token) {
+                navigate('/login');
+                return;
+            }
             await axios.delete(`/${hobbyNumber}/hobby-review/${reviewNumber}`,{
                 headers: {
-                    'Authorization': token // JWT 토큰을 Authorization 헤더에 추가
+                    'Authorization': `Bearer ${token}` // JWT 토큰을 Authorization 헤더에 추가
                 }
             });
             alert("후기가 삭제되었습니다.");
@@ -99,10 +108,14 @@ function HobbyReviewGet() {
     };
 
     try {
-        const token = localStorage.getItem("token");
+        const token = getStoredToken();
+        if (!token) {
+            navigate('/login');
+            return;
+        }
         await axios.put(`/${hobbyNumber}/hobby-review/${reviewNumber}`, reviewData,{
             headers: {
-                'Authorization': token // JWT 토큰을 Authorization 헤더에 추가
+                'Authorization': `Bearer ${token}` // JWT 토큰을 Authorization 헤더에 추가
             }
         });
